move precompile fixture to tests/common
